split profiling
